Fix int32 overflow in warpctc, warprnnt, gru_unit and spectral_norm impl

+++ paddle/phi/kernels/impl/warprnnt_grad_kernel_impl.h
@@ -34,10 +34,10 @@ void WarprnntGradKernel(const Context& dev_ctx,
                         DenseTensor* input_grad) {
   dev_ctx.template Alloc<T>(input_grad);
 
-  int B = warprnntgrad.dims()[0];     // B
-  int Tmax = warprnntgrad.dims()[1];  // Tmax
-  int Umax = warprnntgrad.dims()[2];  // Umax
-  int D = warprnntgrad.dims()[3];     // D
+  int64_t B = warprnntgrad.dims()[0];     // B
+  int64_t Tmax = warprnntgrad.dims()[1];  // Tmax
+  int64_t Umax = warprnntgrad.dims()[2];  // Umax
+  int64_t D = warprnntgrad.dims()[3];     // D
 
   // (B,)
   auto loss_grad_e = EigenTensor<T, 1>::From(loss_grad);
